Collapse semantic caches behind typed accessors
